fix: move DiffEqGPU.jl benchmark to the "lower level API".
Solving the EnsembleProblem with GPUTsit5() and EnsembleProblem(CUDA.CUDABackend()) was stubbornly residing on the CPU only, I couldn't crack it. This version lives on the GPU. I can't quite interpret whether it's transferring memory back to the host - surely it must be to save the output.

This seems like the closest we'll get to an apples-with-apples fair fight. The adaptive solver is slower in this version - usure why.